+++ finagle-core/src/main/scala/com/twitter/finagle/Namer.scala
@@ -72,8 +72,8 @@ object Namer {
 
     private[this] object InetPath {
 
-      private[this] def resolve(host: String, port: Int): Var[Addr] =
-        Resolver.eval(s"inet!$host:$port") match {
+      private[this] def resolve(scheme: String, host: String, port: Int): Var[Addr] =
+        Resolver.eval(s"$scheme!$host:$port") match {
           case Name.Bound(va) => va
           case n: Name.Path =>
             Var.value(
@@ -82,11 +82,16 @@ object Namer {
         }
 
       def unapply(path: Path): Option[(Var[Addr], Path)] = path match {
-        case Path.Utf8("$", "inet", host, IntegerString(port), residual @ _*) =>
-          Some((resolve(host, port), Path.Utf8(residual: _*)))
-        case Path.Utf8("$", "inet", IntegerString(port), residual @ _*) =>
+        case Path.Utf8(
+              "$",
+              scheme @ ("inet" | "fixedinet"),
+              host,
+              IntegerString(port),
+              residual @ _*) =>
+          Some((resolve(scheme, host, port), Path.Utf8(residual: _*)))
+        case Path.Utf8("$", scheme @ ("inet" | "fixedinet"), IntegerString(port), residual @ _*) =>
           // no host provided means localhost
-          Some((resolve("", port), Path.Utf8(residual: _*)))
+          Some((resolve(scheme, "", port), Path.Utf8(residual: _*)))
         case _ => None
       }
     }